Or statements for fanSwitch
Now, this should work. The if - or statement checks to see if any of the off conditions are met. If they are, fanSwitch becomes false. fanSwitch is returned to true every cycle, and only if a switch is tripped does that change. (perhaps makes more sense to default to false instead of true...)